Create library class & use SOLID principles
